permissions fixed for pathfinding. Zoom is still under investigation

+++ src/main/java/controllers/MapOverlay.java
@@ -225,8 +225,8 @@ public class MapOverlay {
 
         // this code sets node's x and y pos to be on the plane holding the graph
         currentPane.getChildren().add(location);
-        location.setLayoutX(nodeX * zoom * widthRatio);
-        location.setLayoutY(nodeY * zoom * heightRatio);
+        location.setLayoutX((int)(nodeX * zoom * widthRatio));
+        location.setLayoutY((int)(nodeY * zoom * heightRatio));
         location.toFront();
 
         if (!enabled) {
